--- a/src/Composer/DependencyResolver/PoolBuilder.php
+++ b/src/Composer/DependencyResolver/PoolBuilder.php
@@ -129,6 +129,7 @@ class PoolBuilder
             // replace means conflict, so if a fixed package replaces a name, no need to load that one, packages would conflict anyways
             foreach ($package->getReplaces() as $link) {
                 $this->loadedPackages[$link->getTarget()] = $link->getConstraint();
+                $this->skippedLoad[$link->getTarget()] = $package->getName();
             }
 
             // TODO in how far can we do the above for conflicts? It's more tricky cause conflicts can be limited to