--- a/tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/package-versions-are-not-loaded-if-not-required-recursive.test
+++ b/tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/package-versions-are-not-loaded-if-not-required-recursive.test
@@ -1,5 +1,5 @@
 --TEST--
-Test irrelevant package versions are not loaded
+Test irrelevant package versions are not loaded recursively
 
 --REQUEST--
 {