… composer.json
… composer.lock
… src/Composer/DependencyResolver/Request.php
… src/Composer/Package/Link.php
… src/Composer/Repository/ArrayRepository.php
… src/Composer/Repository/ComposerRepository.php
… src/Composer/Repository/CompositeRepository.php
… src/Composer/Repository/FilterRepository.php
… src/Composer/Repository/PlatformRepository.php
… src/Composer/Repository/RepositoryInterface.php
… tests/Composer/Test/Autoload/AutoloadGeneratorTest.php
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/alias-priority-conflicting.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/alias-with-reference.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/constraint-expansion-works-with-exact-versions.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/fixed-packages-do-not-load-from-repos.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/fixed-packages-replaced-do-not-load-from-repos.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/multi-repo-replace-partial-update.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/multi-repo-replace.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/must-expand-root-reqs.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/package-versions-are-not-loaded-if-not-required-expansion.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/package-versions-are-not-loaded-if-not-required-simple.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/packages-that-do-not-exist.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update-transitive-deps-no-root-unfix.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update-transitive-deps-unfix.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update-unfixing-with-replacers.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/partial-update.test
… tests/Composer/Test/DependencyResolver/Fixtures/poolbuilder/stability-flags-take-over-minimum-stability-and-filter-packages.test
… tests/Composer/Test/DependencyResolver/PoolBuilderTest.php
… tests/Composer/Test/DependencyResolver/RequestTest.php
… tests/Composer/Test/DependencyResolver/RuleTest.php