This is a follow up of the work I did in https://github.com/composer/composer/pull/8275 and @ausi tried to extend even more in https://github.com/composer/composer/pull/8295.
With the new `PoolBuilder` things have changed quite a lot since those attempts so here's another one.

What happens currently is that early package filtering happens only for root dependencies. So the dependent packages are not filtered at all.
Let's take an example to illustrate: Let's take the package `api-platform/api-pack` in version `^1.2`.

Because `api-platform/api-pack` is a root requirement, all the package versions `<1.2.0.0` are not considered. Their information is downloaded from Packagist but it doesn't make it into the `Pool` of rules that are considered.
However, all the packages that `api-platform/api-pack` requires (which are quite a lot) are not filtered. For these packages, **all** versions are loaded into the dependency resolving process.

This PR tries to improve that by remembering the last constraint of a certain package that was loaded and if things do not match, extend that and load it again. Example:

1. `api-platform/api-pack` requires `doctrine/orm` with `^2.4.5`
2. with this PR, only the versions of `doctrine/orm` that match `^2.4.5` are loaded!
3. if any other package was to require `doctrine/orm` in e.g. `^2.6`, nothing happens because `^2.4.5` also matches `^2.6`.
4. if any other package was to require `doctrine/orm` in e.g. `^2.3`, we turn the original constraint from `^2.4.5` into `^2.3 || ^2.4.5` and mark the package as not loaded so eventually, the missing packages will make it into the resolving process as well :)

Thanks to my previous work on the `MultiConstraint::create()` in `composer/semver` and @jderusse's latest work, we can optimize these constraints quite a bit along the way which should result in even better numbers :)

This optimization should improve all kinds of combinations but it will **massively** improve projects that depend on meta packages such as the `api-platform/api-pack` example.
Here are some tests:

For `"api-platform/api-pack": "^1.2"`:

|  | 1.10.5 | master@56edd5304 | this PR |
|------|----------:|-----------------:|--------:|
| RAM | 2054.08MB | 882.87MB | 97.61MB |
| Time | 64.48s | 27.17s | 3.65s |

For `"contao/manager-bundle": "^4.9"`:

|  | 1.10.5 | master@56edd5304 | this PR |
|------|---------:|-----------------:|---------:|
| RAM | 1999.2MB | 1170.1MB | 198.74MB |
| Time | 137.71s | 66.57s | 10.18s |

I'm not going to comment on the numbers, they should speak for themselves :)

I did not encounter any issues except for one project where it ended in a non-resolvable set of dependencies. But for that project, dependencies are so complex, it's very hard to debug. I guess there's just a use case I've not considered yet. My hope is that maybe someone else will have the same issue with a simpler dependency layout so debugging might become easier ;)