@naderman 2020-04-29
This is looking like a great start. I think this really needs to come with a suite of tests for the pool builder to check which packages are loaded giving a certain set of inputs. Any bug that leads to unnecessary loading too many packages will be undetectable elsewhere since the solver will just pick the right versions anyway. So really the only way to make sure this doesn't break in the future is to add tests for which things should or should not be in the pool in a given scenario.