@Toflar 2020-05-04
This PR is on hold because it cannot be finished until https://github.com/composer/semver/issues/94 is resolved.