@naderman 2020-05-25
Reminder for something to look into later: We take some information from the latest version of a package, e.g. abandoned flag. Need to check this still works if we now only partially load versions per package.