@stof 2020-05-25
@naderman shouldn't Packagist mark all versions as abandoned instead ?