@Seldaek 2020-06-06
Tests are fixed now. Last commit also fixed a couple issues due to extending the loaded packages beyond the root constraints, which makes no sense as the root package requirements always have to be fulfilled. 

So https://github.com/composer/composer/pull/8850/commits/2fa58ccf963ca5f4f0e059d0be238d8f4226e519 now forces it to load the root constraint at most (for unfixed packages, it'll load the least possible, up to a max of the root constraint, for non fixed ones it'll just load the whole root constraint at the beginning).

Last thing I'd like to do is try to optimize the loading of ComposerRepository to avoid it reloading the same packages over and over. Won't have time for this today though. Not sure if more PoolBuilder tests are needed or if we're happy with the current set? @naderman?

@Toflar can you make it run on those projects where you got errors at some point to see if it behaves normally now?