@naderman 2020-06-07
@Seldaek there were definitely still tests missing, did you consider all the situations regarding replacers matching the root requirements when making the change you described (haven't looked at code changes yet)