@Toflar 2020-06-08
> @Toflar can you make it run on those projects where you got errors at some point to see if it behaves normally now?

Yep, works normally now! All three runs return the same output (I mean, would update to the same package combination). Here are the profile results on a pretty huge project (248 packages):

| Branch                  | Results                                                  |
|-------------------------|----------------------------------------------------------|
| 1.10                    | Memory usage: 562.19MiB (peak: 2486.33MiB), time: 69.29s |
| master@9d5051f          | Memory usage: 91.17MiB (peak: 1626.48MiB), time: 44.13s  |
| filter-packages@2fa58cc | Memory usage: 63.3MiB (peak: 1033.55MiB), time: 28.87s    |

🎉