@Seldaek 2020-07-17
Ok all tests are back to green, rebased on latest master, ready to merge IMO..

It would be great if we can get more people to test this on various projects though to make sure it does not regress in any major way. From our limited testing it seems to have benefits in some cases but not all (which is fine).

## You can help test this PR like this:

- `cd` into a project you want to try out with
- Download the phar file for this branch: `wget https://seld.be/composer-test.phar`
- Run `composer self-update --snapshot` to get the latest 2.0 dev build
- Run `composer update --dry-run` to prime the cache
- Run `COMPOSER_DISABLE_NETWORK=1 composer update -v --dry-run --profile` to run a profile without network jitter, with current master branch
- Run `COMPOSER_DISABLE_NETWORK=1 php composer-test.phar update -v --dry-run --profile` to run another profile with the optimizations in this branch
- Make sure the results are the same 👀 e.g. the `Lock file operations: 3 installs, 61 updates, 0 removals` lines in both runs should have similar numbers.
- Report results here 📝 We're mostly interested in the last lines of each profile run, e.g. `Memory usage: 90.53MiB (peak: 1213.05MiB), time: 19.38s`, but if you spot a discrepancy in what gets updated please do tell as that would be bad.
- `Unsubscribe` from notifications on the top right of this page if you don't want to get spammed with everyone else's results 😃

If you're interested in more details, you can look at the `Analyzed x packages/rules` lines which show up somewhere on top to see how many packages and inter-package dependency rules were saved by this PR.

Thanks! 🚀