@Bacteries 2020-07-17
First : 
> Memory usage: 49.57MiB (peak: 804.99MiB), time: 14.81s

Second : 
> Memory usage: 71.03MiB (peak: 804.9MiB), time: 14.68s

And for each it was the same changes : _Package operations: 3 installs, 31 updates, 5 removals_