@rivimey 2020-07-17
Just tried against a drupal 8 setup in need of updates. The composer.json includes "cwegans/composer-patches" which fails to resolve for composer 2.

Running:
    COMPOSER_DISABLE_NETWORK=1 composer update -v --dry-run --profile
uses my 1.0-stable composer:

`[10.2MiB/0.01s] Loading composer repositories with package information
[10.7MiB/0.58s] Updating dependencies (including require-dev)
[545.1MiB/17.64s] Dependency resolution completed in 0.450 seconds
[363.0MiB/17.81s] Analyzed 14313 packages to resolve dependencies
[363.0MiB/17.81s] Analyzed 317223 rules to resolve dependencies
[364.9MiB/17.86s] Dependency resolution completed in 0.001 seconds
[360.4MiB/17.86s] 48 packages you are using are looking for funding.
[360.4MiB/17.86s] Use the composer fund command to find out more!
[360.4MiB/17.87s] Memory usage: 360.43MiB (peak: 545.36MiB), time: 17.87s
`

Then running with composer-test.phar:

`[12.2MiB/0.01s] Loading composer repositories with package information
[12.4MiB/0.01s] Updating dependencies
[12.6MiB/0.22s] https://packages.drupal.org/8 could not be fully loaded (Network disabled, request canceled: https://packages.drupal.org/8/packages.json), package information was loaded from the local cache and may be out of date
[20.9MiB/0.47s] https://repo.packagist.org could not be fully loaded (Network disabled, request canceled: https://repo.packagist.org/packages.json), package information was loaded from the local cache and may be out of date
[468.9MiB/4.46s] Dependency resolution completed in 0.001 seconds
[469.0MiB/4.46s] Your requirements could not be resolved to an installable set of packages.
[469.0MiB/4.46s]
  Problem 1
    - drupal-composer/drupal-scaffold[2.5.0, 2.5.1, 2.5.2, 2.5.3, 2.5.4, 2.6.0, 2.6.1] require composer-plugin-api ^1.0.0 -> found composer-plugin-api[2.0.0] but it does not match the constraint.
    - Root composer.json requires drupal-composer/drupal-scaffold ^2.5 -> satisfiable by drupal-composer/drupal-scaffold[2.5.0, 2.5.1, 2.5.2, 2.5.3, 2.5.4, 2.6.0, 2.6.1].
`