@1ma 2020-07-17
```
composer 2.0-dev (2.0-dev+c3028c02d95bd68a0a1bcc035d7d75561c916d82) 2020-07-17 11:18:02
Package operations: 2 installs, 18 updates, 0 removals
Memory usage: 36.62MiB (peak: 283.48MiB), time: 7.81s

composer-test.phar (commit 071350286fb196722728511aecd00e1defa826c7)
Package operations: 2 installs, 18 updates, 0 removals (same)
Memory usage: 37.95MiB (peak: 283.52MiB), time: 7.76s
```