@navarr 2020-07-17
snapshot: 
```
Package operations: 1 install, 10 updates, 76 removals
Memory usage: 57.89MiB (peak: 631.16MiB), time: 17.33s
```

test:
```
Package operations: 1 install, 10 updates, 76 removals
Memory usage: 81.69MiB (peak: 598.12MiB), time: 17.75s
```

Use case: Magento installation with replacements and path repository usage