@deresh 2020-07-17
snapshot

`Package operations: 5 installs, 74 updates, 1 removal`
`[54.6MiB/8.14s] Memory usage: 54.6MiB (peak: 888.81MiB), time: 8.14s`

test

`Package operations: 5 installs, 74 updates, 1 removal`
`[55.9MiB/7.86s] Memory usage: 55.87MiB (peak: 847.81MiB), time: 7.86s`

case: custom symfony 5 app